// src/components/Reports/GoalCalculator/CalculatorSettings/Steps/InformationStep/InformationStepForm/enums.ts
export enum Role {
  Office = 'Office',
  Field = 'Field',
}

export enum BenefitsPlan {
  Select = 'Select',
  Plus = 'Plus',
  Base = 'Base',
  Minimum = 'Minimum',
}

export enum FamilySize {
  SingleOrSpouseNotStaff = 'SingleOrSpouseNotStaff',
  MarriedNoChildren = 'MarriedNoChildren',
  Sosa1Dependent = 'Sosa1Dependent',
  Sosa2Or3Dependents = 'Sosa2Or3Dependents',
  Sosa4OrMoreDependents = 'Sosa4OrMoreDependents',
  Married1Or2Children = 'Married1Or2Children',
  Married3OrMoreChildren = 'Married3OrMoreChildren',
}

export enum Tenure {
  OneToFour = '1-4',
  FiveToNine = '5-9',
  TenToFourteen = '10-14',
  FifteenToNineteen = '15-19',
  TwentyToTwentyFour = '20-24',
  TwentyFiveToTwentyNine = '25-29',
  ThirtyToThirtyFour = '30-34',
  ThirtyFiveToThirtyNine = '35-39',
  FortyToFortyFour = '40-44',
  FortyFivePlus = '45+',
}

export enum Age {
  Under30 = 'Under30',
  ThirtyToThirtyFour = '30-34',
  ThirtyFiveToThirtyNine = '35-39',
  Over40 = 'Over40',
}